[cmdline args] Add support for directories
Rather than requiring the user to pass a quoted wildcard to lnav,
we should just support scanning whole directories and reading
them in.  This change adds support for that by detecting the
directory and adding the wildcard to the file list automatically.
Any non-file entries in the given directory are skipped instead
of walking the whole tree, which seems like a reasonable default.