[build] Fix the 'readline' ac macro.
The macro does not track the include directory when it is passed in
using the '--with-readline=/foo/bar' invocation of '--with-readline'
flag.

This change should fix the problem. I've tried to keep the functionality
exactly the same as before and converted the script to pure M4sh, which
should produce a more portable configure script.