[build] Make sure readline has the necessary symbols.
On FreeBSD variant, readline functionality is provided by editline which
is almost compatible but not quite. We could do a specific check for
editline but if editline adds the functionality in the future, this
check will still continue to fail. Instead look for one of the symbols
currently missing in editline as a sanity check.