[syslog_log] Case-insensitive search for level strings
While looking through syslog messages on OS X, came across plenty of
error/warning messages where the level was in all-caps and as a result
were not recognized. There was also a lot of false positives due to the
word error turning up as part of a longer string.

This change:
- Changes the log level regex to do a case-insensitive search.
- Ognores the string 'error', if it is part of a longer string.