fix compile on SUSE
If we have both 64-bit and 32-bit versions of ncurses installed,
currently the 32-bit libraries in /lib get found first, so gcc fails to
link against them, and ./configure ends up thinking that we're missing
ncurses libraries.

So scan /lib64 before /lib.  There might be a better fix, but hopefully
it will avoid breaking 32-bit distributions and distributions which have
64-bit libraries in /lib.