[ui] rework the default colors change
The default colors change resulted in some parts of the lnav
UI being colored incorrectly.  I've reworked it a bit so that
the ansi color pairs are preserved and the default colors are
only used with the color pairs for roles.  I've also replaced
references to ansi colors with roles in a few places to make
things work consistently.