linebuffer: Ensure get_range returns in-range data
In get_range if the requested offset is not in the cache, we could return
a very large integer in avail_out.  This is caused by converting a
negative value into an unsigned (size_t).  It seems like we should
`require` this value to be in-range, but it seems like several callers
already call us without their offset already in range.  They seem to
accept a zero-length to mean we don't have their data, though.

Ensure that the returned avail_out parameter doesn't try to go negative.
If the requested offset is not cached, return a zero-length width so the
caller doesn't think we have anything useful for them.