Add another filename mode to show file path
Add a separate BASENAME bit to distinguish from the FILENAME bit.
When BASENAME is active show the base filename in the filename column.
When FILENAME is active show the whole path for each filename.  Cycle
through the three modes with toggle_filename() still. Note that we are
never in both FILENAME and BASENAME mode.

The user can cycle through the three modes by pressing the hotkey an
extra time.  First press shows the basename, second press shows the full
path (filename), and the third press hides the filename column again.