[view_curses] Fix for fg colour and reverse attribute for separator bars
Elements of type VC_GRAPHIC used as column separators, when used
along with background coloured bars to express numeric column values,
maintain alternate-line reversing of colours when inside the bar range
and take the foreground color of the last coloured bar on the line even
though are not placed in the coloured range of characters.
This commit fixes this.