[views_vtab] add a movement column to lnav_views and other stuff (related to #1088)
I've plumbed the "movement" control through the lnav_views table
so that it can be controlled through the keymap definition instead
of hard-coded in hotkeys.cc.  (Trying to move most stuff out of
there...)

Removed listview_curses::lv_displayed_rows since it seemed to
cause some glitchy behavior when moving to the last line.

Call invoke_scroll() when changing the selection and change the
bottom status to reflect the selected line instead of just the
top.  Also, changing the selectability should set "needs update".

Change more commands to use the selection instead of the top.