strnatcmp.c: ensure correct value range for isdigit() argument.
The valid values to pass to `isdigit()` is the values represented
by `unsigned char` and the value of EOF (usually -1).  Other values
such as the other negative `signed char` values may invoke undefined
behaviour.

Fix this by casting the argument to `isdigit()` to `unsigned char`.

Found by building on NetBSD/macppc with -Wchar-subscripts turned on.