ArenaAlloc/arenaallocimpl.h: avoid shift overflow on 32-bit hosts.
Use the constants defined by <stdint.h> to avoid right-shift by 32
on a 32-bit host by comparing SIZE_MAX to UINT32_MAX, since `value`
is a size_t.

Found by building on NetBSD/macppc with -Wshift-count-overflow
(which is default on in the pkgsrc setup, which this is from).

./third-party/ArenaAlloc/arenaallocimpl.h:111:22: warning: right shift count >=
width of type [-Wshift-count-overflow]
  111 |       value |= value >> 32;
      |                ~~~~~~^~~~~